feat: Add Packages component
This commit adds a new component called Packages to the application. The component renders a view with the text "Packages". This addition enhances the functionality of the application by providing a new screen for displaying packages.

Note: This commit message follows the established convention of using the "feat" type for adding new features.