Refactor ComfirmedListModal component to use presence and friends data
This commit refactors the ComfirmedListModal component to use the presence and friends data retrieved from the server. It replaces the previous hardcoded data with the actual data fetched from the API. The component now uses the PresenceContext to get the presence data and the AuthContext to get the friends data. The presence data is mapped to include additional information such as the profile photo and whether the user is a friend. The friends data is used to determine if a user is a friend or not. This improves the accuracy and relevance of the displayed information in the ComfirmedListModal component.